Rename all images referenced in index.js
